+++ index.js
@@ -1,4 +1,11 @@
-// This is the only way I've been able to copy the ValidityState object.
+/**
+ * extractValidity - Extracts the ValidityState information from a given
+ * object into an object suitable for manipulation.
+ *
+ * @param  {HTMLElement} el A DOM element containing a ValidityState object.
+ * @return {object}         A non-read-only object mimicing the ValidityState
+ *                          object for the given element.
+ */
 function extractValidity(el) {
   const validity = el.validity
 
@@ -35,14 +42,18 @@ function extractValidity(el) {
 }
 
 export default class VueForm {
-  constructor (options = {
-    wasFocusedClass: 'wasFocused',
-    wasSubmittedClass: 'wasSubmitted',
-    noValidate: true
-  }) {
-    this.$noValidate = options.noValidate
-    this.$wasFocusedClass = options.wasFocusedClass
-    this.$wasSubmittedClass = options.wasSubmittedClass
+  constructor (options) {
+    const defaults = {
+      wasFocusedClass: 'wasFocused',
+      wasSubmittedClass: 'wasSubmitted',
+      noValidate: true,
+      required: []
+    }
+    Object.assign(defaults, options)
+    this.$noValidate = defaults.noValidate
+    this.$wasFocusedClass = defaults.wasFocusedClass
+    this.$wasSubmittedClass = defaults.wasSubmittedClass
+    this.$requiredFields = defaults.required
     this.$wasSubmitted = false
     this.$isInvalid = false
     this.$isValid = true
@@ -50,27 +61,33 @@ export default class VueForm {
   }
 
   static install (Vue) {
+
     // v-form directive.
     Vue.directive('form', {
+
+      // Setup the form object when the directive is first bound to the
+      // form element.
       bind (el, { value }) {
         value.$el = el
         value.$el.noValidate = value.$noValidate
 
-        //
+        // Update the forms $wasSubmitted state and apply the appropriate CSS
+        // class when the forms submit event is triggered.
         value.$el.addEventListener('submit', () => {
           value.$wasSubmitted = true
           value.$el.classList.add(value.$wasSubmittedClass)
         })
 
+        // Update the form and child field state and remove any corresponding
+        // CSS classes when the forms reset event is triggered.
         value.$el.addEventListener('reset', () => {
-          // Reset $wasSubmitted property.
           value.$wasSubmitted = false
           value.$el.classList.remove(value.$wasSubmittedClass)
 
           // Reset $wasFocused property and remove the corresponding class
           // from each child node.
           for (const id of Object.keys(value)) {
-            if (id.indexOf('$') === -1) {
+            if (id.indexOf('$') === -1 && value[id].$el) {
               value[id].$wasFocused = false
               value[id].$el.classList.remove(value.$wasFocusedClass)
               Object.assign(value[id], extractValidity(value[id].$el))
@@ -79,33 +96,59 @@ export default class VueForm {
           }
         })
 
+        // Go through each field within the form, set up its state within
+        // the form object, and listen to input or change events to keep its
+        // state in sync.
         for (const $el of el.querySelectorAll('input, textarea, select')) {
-          //
+
+          // Only work with elements that belong to the form, have the ability
+          // to be validated, and have and id or name property.
           if ($el.form === el && $el.willValidate && $el.hasAttribute('id')) {
-            //
+
+            // Create the field object and extract its validity state.
             const field = Object.assign({ $el }, extractValidity($el))
             const id = $el.getAttribute('id')
             Vue.set(value, id, field)
             value.$updateFormValidity(id)
+            value.$updateNamedValidity($el, Vue)
 
             // Add wasFocused class to element when focus event is triggered.
             $el.addEventListener('focus', ({ target }) => {
-              const id = $el.getAttribute('id')
               value[id].$wasFocused = true
               target.classList.add(value.$wasFocusedClass)
             })
 
-            $el.addEventListener('input', ({ target }) => {
-              const id = target.getAttribute('id')
+            // On change or input events, update the field and form validity
+            // state.
+            const type = $el.getAttribute('type')
+            const isCheckable = ['radio', 'checkbox'].indexOf(type) !== -1
+            const eventType = isCheckable ? 'change' : 'input'
+            $el.addEventListener(eventType, ({ target }) => {
               Object.assign(value[id], extractValidity(target))
               value.$updateFormValidity(id)
+              value.$updateNamedValidity(target, Vue)
             })
+
           }
+
         }
       }
     })
+
   }
 
+  /**
+   * setCustomValidity - A wrapper for HTML5s setCustomValidity function so that
+   * the end user can trigger a custom error without an error message, the
+   * custom error message is accessible through the form object, and the overall
+   * form validity is updated.
+   *
+   * @param {string}          field   The identifier for the field you wish to
+   *                                  set the validity for.
+   * @param {boolean|string}  invalid Whether the field is invalid (true), or
+   *                                  not (false), or the custom error message
+   *                                  for an invalid field.
+   */
   $setCustomValidity (field, invalid) {
     if (this[field]) {
       const isBoolean = typeof invalid === 'boolean'
@@ -126,6 +169,15 @@ export default class VueForm {
     }
   }
 
+  /**
+   * updateFormValidity - Updates the overall validity of the form based on the
+   * existing validity state of its fields and the updated validity state of
+   * the given field.
+   *
+   * @param {string} field The identifier for the field whose validity state
+   *                       has updated and has consequently triggered the update
+   *                       of the overall forms validity.
+   */
   $updateFormValidity (field) {
     const index = this.$invalidFields.indexOf(field)
     if (this[field].valid && index !== -1) {
@@ -139,5 +191,40 @@ export default class VueForm {
       this.$isInvalid = true
       this.$invalidFields.push(field)
     }
+  }
+
+
+  /**
+   * $updateNamedValidity - For the use case of requiring a value for a set of
+   * checkboxes or radio buttons with the same name, VueForm provides the
+   * validity state of the overall group using the name as the identifier. This
+   * function updates this validity state.
+   *
+   * @param {HTMLElement} el  The DOM element that may trigger an update to the
+ *                            validity of the named group.
+   * @param {Vue}         Vue The Vue.js instance given when this plugin is
+   *                          installed.
+   */
+  $updateNamedValidity (el, Vue) {
+
+    // Check if the element has a name
+    if (el.hasAttribute('name')) {
+      const name = el.getAttribute('name')
+
+      // Check if the named group was marked as required.
+      if (this.$requiredFields.indexOf(name) !== -1) {
+
+        // Set the validity state of the named group.
+        const valid = new FormData(this.$el).has(name)
+        const validity = { valid, valueMissing: !valid  }
+        if (this[name]) {
+          Object.assign(this[name], validity)
+        } else {
+          Vue.set(this, name, validity)
+        }
+
+      }
+    }
+
   }
 }