+++ src/components/SiteFooter.vue
@@ -10,3 +10,9 @@
     All rights reserved.
   </footer>
 </template>
+
+<script>
+  export default {
+    name: 'SiteFooter'
+  }
+</script>