+++ test/e2e/specs/contact.js
@@ -1,17 +1,14 @@
-// For authoring Nightwatch tests, see
-// http://nightwatchjs.org/guide#usage
-
 module.exports = {
-  'Home Validation Failure Test': browser => {
-    console.log(browser.page.home())
-    browser.page.home().navigate()
+  'Contact Form Validation Failure Test': browser => {
+    console.log(browser.page.contact())
+    browser.page.contact().navigate()
       .waitForElementVisible('@contactForm', 10000)
       .submitForm('@contactForm')
       .assert.containsText('@contactForm', 'Name is required.')
       .assert.containsText('@contactForm', 'Email is required.')
       .assert.containsText('@contactForm', 'Phone is required.')
       .assert.containsText('@contactForm', 'Description is required.')
-      .assert.urlEquals(browser.page.home().url())
+      .assert.urlEquals(browser.page.contact().url())
       .setValue('@email', 'notanaccount')
       .assert.containsText('@contactForm', 'Please enter a valid email address.')
       .api.end()